python-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
Pillow==11.0.0
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.2
imagehash==4.3.1
chardet==5.2.0
python-dotenv==1.0.1
torch==2.5.1+cpu
sentence-transformers==2.2.2
chromadb==0.4.24
keybert==0.8.3
transformers==4.45.1
langchain==0.2.5
openai==1.50.2
httpx==0.27.2
numpy==1.26.4

